requirements.txt
customtkinter>=5.2.0
pywin32>=306
pillow>=10.0.0  # Required for customtkinter
pyinstaller>=6.0.0
winshell>=0.6; platform_system == "Windows"
